#36 refactor: UpdateUser 시, 유저 데이터 갱신 방식 변경

+++ domain/src/main/java/io/tuttut/domain/usecase/user/UpdateUserUseCase.kt
@@ -27,14 +27,19 @@ class UpdateUserUseCase @Inject constructor(
             val updateUserRequest = UpdateUserRequest(name, profile)
             authRepository.updateUser(credential.userId, updateUserRequest)
                 .also {
-                    if (currentUser.profile != profile) {
+                    if (currentUser.profile.isGoogleProfile().not() && currentUser.profile != profile) {
                         deleteImageUseCase(currentUser.profile, SaveLocation.USER)
                     }
-                    preferenceRepository.updateCurrentUser(updateUserRequest)
+                    val updatedUser = currentUser.copy(name = name, profile = profile)
+                    preferenceRepository.setCurrentUser(updatedUser)
+                    preferenceRepository.setGardenUser(updatedUser)
                 }
         }
     }
 
     private fun isUserDataChanged(user: User, name: String, imageSource: ImageSource): Boolean =
         user.name != name || user.profile != imageSource
+
+    private fun ImageSource.Remote.isGoogleProfile() =
+        url.contains("googleusercontent")
 }
